refactor: moved CustomHealthRecordingScreen to folder
